commit for class 11/1

+++ src/Geometry2D/Line.java
@@ -2,11 +2,11 @@ package Geometry2D;
 
 public class Line {
 
-    public int x1;
-    public int y1;
-    public int x2;
-    public int y2;
-    public Line (int x1, int y1, int x2, int y2) {
+    public double x1;
+    public double y1;
+    public double x2;
+    public double y2;
+    public Line (double x1, double y1, double x2, double y2) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
